requirements.txt
Django==3.2.6
# DESCOMENTAR PARA UTILIZAR MYSQL mysqlclient==2.1.0
djangorestframework==3.12.4